Split tags into their own model again
