+++ myopenpantry/models/__init__.py
@@ -1,4 +1,5 @@
 from .associations import RecipeIngredients, recipe_tags # noqa
 from .ingredients import Ingredient # noqa
-from .recipes import Recipe, Tag # noqa
+from .recipes import Recipe # noqa
 from .items import Item # noqa
+from .tags import Tag # noqa
