+++ myopenpantry/models/recipes.py
@@ -4,14 +4,6 @@ from myopenpantry.extensions.database import db
 
 from datetime import datetime
 
-# classify a recipe for keyword searches, eg "chicken", "french", "quick"
-class Tag(db.Model):
-    __tablename__ = "tags"
-
-    id = db.Column(sa.Integer, primary_key=True)
-    name = db.Column(sa.Text)
-    recipes = db.relationship('Recipe', secondary='recipe_tags', back_populates="tags")
-
 class Recipe(db.Model):
     __tablename__ = "recipes"
